[refactor] 임시저장 기능 추가로 인한 board 리팩터링
