+++ src/main/java/com/example/demo/domain/board/domain/vo/Status.java
@@ -1,0 +1,6 @@
+package com.example.demo.domain.board.domain.vo;
+
+public enum Status {
+    DRAFT,
+    PUBLISHED
+}
